Added fix for lab 2: Rename Bus -> DoubleBus and Car -> Bus
