+++ WindowsFormsCars/Bus.cs
@@ -1,50 +1,112 @@
 ﻿using System.Drawing;
+using static WindowsFormsCars.DirectionClass;
 
 namespace WindowsFormsCars
 {
-    class Bus : Car
+    class Bus : Vehicle
     {
-        public Color DopColor { private set; get; }
+        /// <summary>
+        /// Ширина автомобиля
+        /// </summary>
+        protected const int carWidth = 140;
+
+        /// <summary>
+        /// Высота автомобиля
+        /// </summary>
+        protected const int carHeight = 100;
 
         /// <summary>
         /// Конструктор
         /// </summary>
         /// <param name="maxSpeed">Максимальная скорость</param>
         /// <param name="weight">Вес автомобиля</param>
-        /// <param name="mainColor">Основной цвет</param>
-        /// <param name="dopColor">Дополнительный цвет</param>
-        public Bus(int maxSpeed, float weight, Color mainColor, Color dopColor) : base(maxSpeed, weight, mainColor)
+        /// <param name="mainColor">Основной цвет автомобиля</param>
+        public Bus(int maxSpeed, float weight, Color mainColor)
         {
-            DopColor = dopColor;
+            MaxSpeed = maxSpeed;
+            Weight = weight;
+            MainColor = mainColor;
         }
 
         /// <summary>
-        /// Метод отрисовки автобуса
+        /// Метод перемещения автомобиля
+        /// </summary>
+        /// <param name="direction">Направление перемещения</param>
+        public override void MoveTransport(Direction direction)
+        {
+            float step = MaxSpeed * 100 / Weight;
+            switch (direction)
+            {
+                case Direction.Left:
+                    {
+                        if (_startPosX - step > 0)
+                        {
+                            _startPosX -= step;
+                        }
+                        break;
+                    }
+
+                case Direction.Right:
+                    {
+                        if (_startPosX + step + carWidth < _pictureWidth)
+                        {
+                            _startPosX += step;
+                        }
+                        break;
+                    }
+
+                case Direction.Up:
+                    {
+                        if (_startPosY - step > 0)
+                        {
+                            _startPosY -= step;
+                        }
+                        break;
+                    }
+
+                case Direction.Down:
+                    {
+                        if (_startPosY + step + carHeight < _pictureHeight)
+                        {
+                            _startPosY += step;
+                        }
+                        break;
+                    }
+            }
+        }
+
+        /// <summary>
+        /// Метод отрисовки автомобиля
         /// </summary>
         public override void DrawCar(Graphics g)
         {
-            base.DrawCar(g);
-
             Pen pen = new Pen(Color.Black);
+
+            // Корпус автобуса
+            Brush brush = new SolidBrush(MainColor);
+            g.FillRectangle(brush, _startPosX, _startPosY, carWidth, carHeight);
+
+            // Колеса
+            Brush brushBlack = new SolidBrush(Color.Black);
+            g.FillEllipse(brushBlack, _startPosX + 15, _startPosY + carHeight - 15, 30, 30);
+            g.FillEllipse(brushBlack, _startPosX + carWidth - 55, _startPosY + carHeight - 15, 30, 30);
+
+            g.FillEllipse(brush, _startPosX + 20, _startPosY + carHeight - 10, 20, 20);
+            g.FillEllipse(brush, _startPosX + carWidth - 50, _startPosY + carHeight - 10, 20, 20);
 
             //Дверь
             g.DrawRectangle(pen, _startPosX + carWidth - 23, _startPosY + carHeight - 30, 12, 25);
 
             // Лобовое стекло
-            Brush brushBlue = new SolidBrush(DopColor);
+            Brush brushBlue = new SolidBrush(Color.SkyBlue);
             g.FillRectangle(brushBlue, _startPosX + carWidth - 25, _startPosY + 18, 26, 50);
 
             // Стекла
-            g.FillRectangle(brushBlue, _startPosX - 1, _startPosY + 18, 30, 20);
-            g.FillRectangle(brushBlue, _startPosX + 32, _startPosY + 18, 30, 20);
-            g.FillRectangle(brushBlue, _startPosX + 65, _startPosY + 18, 45, 20);
-
             g.FillRectangle(brushBlue, _startPosX - 1, _startPosY + 45, 30, 20);
             g.FillRectangle(brushBlue, _startPosX + 32, _startPosY + 45, 30, 20);
             g.FillRectangle(brushBlue, _startPosX + 65, _startPosY + 45, 35, 20);
 
             // Габариты
-            Brush brushBlack = new SolidBrush(Color.Black);
             Brush brushYellow = new SolidBrush(Color.Yellow);
             g.FillRectangle(brushBlack, _startPosX - 1, _startPosY + 70, 10, 10);
             g.FillRectangle(brushYellow, _startPosX - 1, _startPosY + 82, 10, 10);
