+++ WindowsFormsCars/WindowsFormsCars.csproj
@@ -45,7 +45,7 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="Car.cs" />
+    <Compile Include="Bus.cs" />
     <Compile Include="DirectionClass.cs" />
     <Compile Include="FormCar.cs">
       <SubType>Form</SubType>
@@ -56,7 +56,7 @@
     <Compile Include="ITransport.cs" />
     <Compile Include="Program.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
-    <Compile Include="Bus.cs" />
+    <Compile Include="DoubleBus.cs" />
     <Compile Include="Vehicle.cs" />
     <EmbeddedResource Include="FormCar.resx">
       <DependentUpon>FormCar.cs</DependentUpon>
